Use flake8 :restroom:
