+++ requirements.txt
@@ -16,6 +16,6 @@ PyMySQL==0.8.0
 six==1.11.0
 ### Testing
 click==6.7
-pep8==1.7.0
+flake8==3.5.0
 docker==3.1.1
 pytest==3.4.2
