+++ templates/post/view.html
@@ -28,7 +28,7 @@
       <div class="post-header">
         <a class="text bold color theme" itemprop="headline">{{ item.title }}</a>
       </div>
-      <div id="main" class="post-text text justify" itemprop="articleBody">
+      <div id="main" class="post-text text" itemprop="articleBody">
         {{ item.post_text }}
       </div>
